refactor: clean up main.rs
