+++ src/pages/session_week.rs
@@ -4,7 +4,9 @@ use maud::html;
 use serde::Deserialize;
 use urlencoding::encode;
 
-use crate::{auth::is_authenticated, components::page::page, get_redirect_response, AppState};
+use crate::{
+    auth::is_authenticated, components::page::page, util::get_redirect_response, AppState,
+};
 
 #[derive(Deserialize)]
 pub struct WeekForm {
